backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
shapely==2.0.6
supabase==2.4.6
httpx==0.27.2
geopandas==1.0.1
pyproj==3.7.0
rasterio==1.4.3
scikit-image==0.24.0
reportlab==4.2.2
earthengine-api==0.1.405
python-dotenv==1.0.1
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
joblib==1.4.2
beautifulsoup4==4.12.3
pdfplumber==0.11.0
lxml==5.2.2

rtree==1.3.0
python-multipart==0.0.9
